Add Coveralls step into Portal CI (#595)
## 🛠 Changes

* Adds a new `npm run test:ci` npm script which writes a coverage report to `./portal/coverage`
* Update .gitignore to ignore coverage reports
* Update the Portal CI unit test step to run `npm run test:ci`
* Add a Portal CI step to upload the coverage report to Coveralls

## ℹ️ Context for reviewers

This will make it so that PRs report code coverage to Coveralls and Coveralls will add comments to PRs with code coverage details. There may be some more tweaks in future PRs. I tested this out in Coveralls and it does upload the report and show coverage details in the Coveralls UI. However, it didn't add a coverage comment to my PRs. I suspect it needs to have coverage data on `master` before it will add PR comments.

## 👀 Demo

<img width="516" alt="Screen Shot 2020-07-06 at 9 55 43 AM" src="https://user-images.githubusercontent.com/9039672/86603587-2a97f480-bf72-11ea-9f2e-176b7ed33eb2.png">

<img width="1095" alt="Screen Shot 2020-07-06 at 10 18 29 AM" src="https://user-images.githubusercontent.com/9039672/86603590-2cfa4e80-bf72-11ea-9737-5a9e7a87ba2c.png">