CP-607 Add other leave models (#598)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-607

## 🛠 Changes

* Adds three new models: EmployerBenefit, OtherIncome, and PreviousLeave
* Added fields to the Claim model for the three other leave types. They each have a default value of `[]`

## ℹ️ Context for reviewers

These models will be used to collect other leave information from applicants. For now we're expecting these to be stored in the Application but that may change after pilot 2. This is what we're expecting the API to implement but it's not set in stone.

I imagine that we will use these models to set the other leave fields on the Claim but I don't expect us to implement collections for these models because there aren't good keys to use.