CP-606: Workaround to display checklist step states (#625)
## 🎫 Ticket

https://lwd.atlassian.net/browse/cp-606

## 🛠 Changes

- Updated `Step.isComplete` to check whether all fields have a value
- Fixed `fieldHasValue` so that it identifies `false` as a valid value
- Fixed `updateClaim` workaround so that nested fields aren't removed when objects are merged
- Fixed Step 1's checklist description to include correct 2nd list item content
- Fixed exported field name for `continous_leave_periods`
- Exclude `middle_name` and intermittent leave fields from the lists of exported fields so the checklist can show as "Completed"

## ℹ️ Context for reviewers

When validation is implemented, we'll check whether any validation errors exist in order to identify whether a step is complete, however in the meantime we will assume that all exported fields are required. This means that conditionally displayed fields are problematic, since a user who never is displayed those fields will never see a "completed" checklist. To support pilot 2, we can exclude those conditional fields from a page's exported `fields` list and create a backlog task to add those fields to the list when validation is implemented: https://lwd.atlassian.net/browse/CP-625