API-276: Add application to response (#624)
API-276: Added 'data' attribute to the ApplicationUpdateResponse component. The data attribute will contain the ApplicationResponse with the updated application information. If an error is encountered the data attribute will not be present in the response.

Both the application_update and the application_submit methods were updated to include the 'data' attribute with the updated application object.