API-253 Application DB Model Field (#662)
🎫 Ticket
https://lwd.atlassian.net/browse/API-253

🛠 Changes
in the database: application.middle_initial is now application.middle_name as a nullable field
ApplicationResponse and ApplicationRequest have middle_name as an optional field
the schema in openapi.yaml has the middle_name field
ℹ️ Context for reviewers
The Portal currently has the option for collecting an applicant's middle name. With this PR, middle_name will be able to be stored.