CP-566 Update other leave pages to use enum values, not just keys (#708)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-566

## 🛠 Changes

* Updates the employer benefit details and other incomes details pages to use the EmployerBenefitType values and OtherIncomeType values instead of just the keys
* Fixes the name fo the employer benefit details amount InputText so that it actually saves the value to the Claim.

## ℹ️ Context for reviewers

Using only the keys made it awkward to do testing and made it difficult to track the type values back to the enum where they're defined -- instead of using an object property one had to specify the key as a string. This became more of an issue while adding other leave to the review page so this PR is separating out that work.