API: Disable fsync for local database (#709)
Not forcing the database to wait for the filesystem provides a
significant speed boost to database actions and it's not a big deal
in development if the database loses data during a crash.

A full test suite run on my machine goes from 255s to 78s. Saving
~3 minutes subjectively feels significant when running the suite by
hand. There is minimal-to-no impact on the test suite runtime in the
CI environment.