API-298: Register portal users in FINEOS (#694)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-298

## 🛠 Changes

Added a class in api/services/fineos_actions.py to house the logic to register a portal user with FINEOS. The method takes two arguments, employee SSN and employer FEIN, and invokes two FINEOS APIs. The first API is to find the FINEOS employer identifier by FEIN. This employer identifier is input to the second API to register the employee. 

## ℹ️ Context for reviewers

Three test cases coded to work against the mock FINEOS client. If you want to test locally with the real client you have to uncomment the FINEOS environment variables in the docker configuration file.

Reviewers: @jamesbursa