CP-566 Add Other Leave section to review page (#705)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-566

## 🛠 Changes

* Updates the Review page with a section for other leave and reductions"

## ℹ️ Context for reviewers

It's currently possible for users to get to the review page with empty or incomplete other leave entries so this PR allows and handles that case. If any field of an other leave entry is empty we just won't display that field. But we'll still always display an (possibly empty) entry.

Once we're hooked up to the API and are doing validations we'll ensured to have completed other leave entries on the review page and can remove some of the condition rendering in the helper components.

Here's the [review spreadsheet](https://docs.google.com/spreadsheets/d/1GLMzV1pglfNOEe2fIa-0-4YOenZ-1t6B5IkJrd-Qkrc/edit?ts=5ef297b6#gid=1856733632). I went with the text used in the spreadsheet rather than the text in the Figma design on the ticket.

## 👀 Demo

<img width="785" alt="Screen Shot 2020-07-22 at 5 25 12 PM" src="https://user-images.githubusercontent.com/9039672/88230406-5c7eab80-cc40-11ea-8f96-d1da5e570d42.png">