API-365: Add DB Persistence to user registration and implement review feedback (#751)
In API-298: Register portal users in Fineos DONE the basic registration logic was implemented and a web id was created in the FINEOS product but the resulting web id was not persisted in our database for reuse. This ticket implements the persistence and also implements some of the review comments in the PR for API-298.