API-302: workaround no tax identifier sent by the portal (#766)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-302

## 🛠 Changes

- If no tax_identifier is present in the application, use a fake one

## ℹ️ Context for reviewers

In prod only, the portal is configured not to send the tax identifier to the API at all. This results in an error at submit time:

AttributeError: 'NoneType' object has no attribute 'tax_identifier' method=POST path=/v1/applications/3eb7d7a0-8d18-4fe4-b17d-918c1aa27dda/submit_application

As prod is not connected to a FINEOS instance, add code to use a fake ssn for now. This will allow the `send_to_fineos` code to be executed against the `MockFINEOSClient` in prod without an error being returned.

This workaround will be removed when we receive approval for PII in prod and the portal starts sending PII to the API in prod.

Reviewers: @sawyerh @lorenyu