API-353 (1/3) Add ecs-tasks module (#722)
https://lwd.atlassian.net/browse/API-353

This adds an ecs-tasks module which pulls out the configuration needed for the following ECS task definitions:

- (API) db-migrate-up
- (API) create-db-users

This module is dependent on the `api` module being deployed for a particular environment, as it relies on the RDS instance that is created. Otherwise, the following resources were previously shared between the API application and ECS tasks and are now separated:

- The security group. The new SG is the same as the API SG, but without any ingress rules.
- The task execution IAM role. The new role is the same as the API role, but with access to different cloudwatch log groups.

## ℹ️ Context for reviewers

We are breaking out the migration task definition so that it can be updated and run before the API deployment.