API-311: Add API for RMV VendorLicenseInquiry requests (#666)
https://lwd.atlassian.net/browse/API-311

In #590, we added an API client for the RMV API that was able to pull down the WSDL and generate the client. This PR adds onto that functionality by allowing the client to make requests to the VendorLicenseInquiry endpoint using the tentative format described in Confluence.

## ℹ️ Context for reviewers

In order to test this more easily without monkeypatching, I split the zeep service layer off into an `ApiCaller` interface/abstract class. 

* The logic for parsing the client certificate and pulling the WSDL is moved into the real Zeep API caller
* A new MockZeepCaller is created for testing route-specific logic (for VendorLicenseInquiry). This is useful because we don't have to deal with WSDLs, certificates, or any request monkeypatching 😁 

The WSDL does not provide details on validation or required fields, but I've encoded the known requirements in the pydantic models and route implementation. Notably, the DOB field must be in `YYYYMMDD` instead of `YYYY-MM-DD`.