Add API team as reviewer on Claim and User portal models (#769)
## 🛠 Changes

-  Tag API engineers on changes to Portal data models (`Claim` and `User`)

## ℹ️ Context for reviewers

There are concerns from both teams that discovering problems at the code review stage is too late, and that there needs to be other things we do earlier on, so this can't be something that's done in isolation. But for now it can be something we try out in addition to other things we are doing.