API: Dynamically create RMV cert in test fixture (#782)
Instead of a static certificate for use in the tests, have the fixture generate
one for each test session. This means one less impenetrable file sitting around
needing to be maintained and having some code to generate certificates makes it
easier to test variations if desired (expired certificates, ones not valid until
the future, passwords, require a specific org be listed, etc.).

Add a `tests/helpers/` directory to hold helper code that is specific to the
tests. Occasionally there are things that are only really useful in the context
of the tests, that don't have a more general use to be included in the `massgov`
package.

Certificate generation code borrowed heavily from
https://stackoverflow.com/a/60804101