API-293: Add pregnant or recent birth flag to Application model. (#783)
Added boolean flag pregnant_or_recent_birth to the database model Application as well as to the ApplicationFactory so it will be included in the request and response bodies under the leave details sub-object.

Added a database test for the attribute and fixed this test class to use the test db connection. Also added a test case to the test_application class.