API-355: Fix Employee count in Eligibility Feed generation (#734)
The WagesAndContributions table determines which Employees worked for
which Employers and when, so we use it to get a list of Employees for a
given Employer.

In the real world there will almost always be multiple
WagesAndContributions records for a given Employee-Employer pair (e.g.,
someone works at a place for more than one quarter).

Since for the Eligibility Feed we only care about if the Employee
has *ever* worked for a given Employer and do not need a list of which
specific time periods, we need to filter the results to just the
distinct Employee IDs.

Previously, we were counting the number of WagesAndContributions records
for the Employee-Employer pair, which isn't what we want. This impacted
the logging and the datablock header for the expected number of records
in the export.

This did not impact the data itself as SQLAlchemy quietly deduplicates
results based on primary key when fetching complete models.

The implemented approach introduces a subquery to perform the count.
It's possible to perform the distinct count directly, which may be more
performant, but also may be less readable/maintainable. Since this code
is likely to undergo revisions in the near term as tracking which
Employee records need included in an Eligibility Feed export get
implemented, which will change how we do all these queries, for now,
go with something that reads a little better and we'll optimize the
queries once we know what the final data layout will be.

https://lwd.atlassian.net/browse/API-355