API-325 (experiment): Run docker commands outside of shell context (#827)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-325

## 🛠 Changes

This makes a small change to the way that we run the API from docker.

As noted in [this blog post](https://husobee.github.io/golang/ecs/2016/05/19/ecs-graceful-go-shutdown.html) and also [this one](https://www.ctl.io/developers/blog/post/gracefully-stopping-docker-containers/), when running without the array syntax for `CMD`, the application will start through a parent `/bin/sh` script. This will not broadcast signals like SIGTERM and SIGKILL to the child process. 

Instead, with the array syntax, the docker container will execute our command directly without a middleware shell. This means that signals are sent directly to the API. The first blog post theorizes that this will speed up the graceful shutdown of containers on ECS, by allowing the ECS agent to successfully stop the containers without waiting the full timeout to run a SIGKILL.