Persist absence case id and notification case id to Application table. (#855)
Persisted the absence case id and the notification case id to the Application model. The attributes already existed in the Application table. Work here was to capture the values returned by the startAbsence API and store them into the corresponding attributes in Application. All code is in fineos_actions.py

Added a test case in test_fineos_actions.py to confirm attributes are properly saved in the database.