API: Updates to run Spectral as a part of API linting (#865)
A recent commit[1] introduced the Spectral tool for linting the APIs OpenAPI
spec. Spectral is a Node.js package, the first for the API component itself, and
so node needs to be made available in all API development environments.

Thus we need to add node to the Docker environment. Additionally, generalize the
`PY_RUN_CMD` infrastructure to support a `RUN_CMD` that can be used in the
Makefile to run general commands inside the Docker dev environment.

Since node dependencies are installed in the project directory, this change also
means that `make deps` needs to be run as a part of getting dev tooling set up.

[1] 3c3ea77a8e8d627f960b92612a69b2182d553a19