API-278: Support `tax_identifier` field on Application requests/responses (#908)
Summary:
- Add `tax_identifier` to request
- Deprecate the `employee_ssn` field in request, which is for the same
  information as `tax_identifier`
- Remove `tax_identifier_last4` from response
- Add `tax_identifier` to response, the value of which is masked
- Light refactoring around where `tax_identifier_last4` property lives

The end-to-end tests still have a reference to `tax_identifier_last4` in a
data-structure that seems to be used to create/update an Application. Since
`tax_identifier_last4` wasn't an existing supported field in an Application
request, these changes have not updated it to `tax_identifier` which would
actually set a field on the Application. Stated differently, decided to keep the
behavior of the end-to-end tests the same (i.e., sending information that is
ignored), rather than "fixing" them, changing their behavior.

https://lwd.atlassian.net/browse/API-278