API-435 : Application Submit endpoint to update submitted_time (#914)
Setting submitted time to current time instead of completed time in the application_submits method. Test cases for endpoint where also refactored as needed..

With new progressive app design part 1 of the application will submit to fineos calling the application_submits endpoint. This should mark the application as submitted. When part 2 of the Application is done a new endpoint will be invoked which will set the completed_time attribute to the current time.