CP-762: Add eslint rule to require ticket references in TODO comments (#689)
## ℹ️ Context for reviewers

This is a proposal to include [a new lint rule](https://github.com/sawyerh/eslint-plugin-todo-plz/blob/master/docs/rules/ticket-ref.md) to require the inclusion of a Jira ticket alongside any TODO comment that gets committed to the main branch.

For example, the following would result in a linter error, blocking a PR from being merged:

```js
// TODO: Save these fields to the API
```

The following would be accepted:

```js
// TODO (CP-123): Save these fields to the API
```

### Why 🧐 

The Portal codebase currently had **66** TODO comments. Of those, **27** (40%) didn't reference a ticket. 

We want to avoid losing track of work that we haven't completed, and unintentionally ship an application that doesn't work end-to-end. Making sure that we have tickets in our backlog is one major way we can avoid losing track of this work. In addition, having ticket references within TODO comments provides an engineer a way to learn additional context, and potentially learn that the work has already been completed, but the TODO got left behind in the code by accident.

## 🛠 Changes

- Adds a new ESLint rule to require the inclusion of a ticket number alongside `TODO` comments
- Added Jira tickets for any TODO's that didn't have a ticket reference
- Update TODO comments to follow a consistent format for ticket references
- Removed obsolete code that had lingering or inaccurate TODO comments