CP-897: Remove leave_period arrays entirely when conditional content is hidden on duration page (#935)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-897

## 🛠 Changes

- Removes entire leave period array (instead of the item at the 0 index) when fields are hidden

## ℹ️ Context

There may be instances where the API expects an array but there can only ever be one object. We would then want to treat the array effectively as an object. In that case, if we should add and remove entire array rather than adding and removing just the 0 indexed object.