CP-846: User is prevented from entering an unacceptable date of birth (#890)
🎫 Ticket
https://lwd.atlassian.net/browse/846

🛠 Changes
 Validation method on ApplicationRequestBody added for date_of_birth field
 Tests for test_application.py updated
ℹ️ Context for reviewers
This is for API work only. The idea is to use the existing validation exceptions used by the OpenAPI validators to return an error response when the date of birth field is invalid.

The data field is left empty for the error response, following the guidance provided in https://lwd.atlassian.net/wiki/spaces/API/pages/295338111/API+Response+Format.