API-92: Run API application behind Gunicorn (#997)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-92

## 🛠 Changes

Updates the API to run behind Gunicorn. This removes the following warning:

```
mass-pfml-api    |    WARNING: This is a development server. Do not use it in a production deployment.
mass-pfml-api    |    Use a production WSGI server instead.
```

## ℹ️ Context for reviewers

Thank you to @apopma and @aplybeah for taking an initial look at implementing Gunicorn with me a while back. It was expected to be a quick change, but ended up needing some workarounds and rummaging in documentation.

----

Gunicorn is a production WSGI server that runs multiple workers in front of your API for handling requests.