API-339: Fail if Total Wages Are < Unemployment Min (#1037)
The PR both checks whether total wages are less than the unemployment minimum and checks whether total wages are less than 30 times the weekly benefit amount. The functions that perform these checks return a boolean, which can later be used to determine which description the response should return to indicate if and why an employee if financially eligible or not.