API: Disabling scheduled RDS downtime for staging + prod (#1108)
We're required to configure a scheduled downtime period (using an EOTSS-provided opaque tagging system) for all nonproduction RDS instances. Stage is "not nonproduction" in EOTSS' reckoning, and shouldn't be scheduled to go down at all.

This change disables scheduled downtime for staging and prod RDS, and enables the same for test RDS. Downtime duration and frequency haven't changed.