CP: Handle validation within BaseApi class (#1104)
## 🛠 Changes

- Moved validation error handling into `BaseApi` which will now throw a `ValidationError` if an API response body includes a non-empty `errors` array

## ℹ️ Context for reviewers

Prior to this, only the `claimsLogic.update` method handled validation errors. In addition, there was a bug that prevented the BaseApi model from throwing any error if the API response body included an empty `errors` array. With these changes, any new API module should inherit validation error handling as long as it wraps the request in a try/catch, like we've been doing.