END-242: Track biz-sim state to avoid resubmitting claims (#1124)
https://lwd.atlassian.net/browse/END-242

This PR is a refactor of the business simulation CLI code in preparation for our first run. It adds:

Tracking of submission state of individual claims, to avoid resubmitting them if the process should fail at any point.
Better logging of what's actually happening using winston.
Additionally, I switched from minimistto yargs for the CLI tool. Yargs is more robust, and allows us to split commands easily between separate files, which cleaned up a lot of the cruft that was accumulating in our CLI file.