API-134 - adding healthcheck to docker-compose file (#841)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-134

## 🛠 Changes

A docker health check is added to the `mass-pfml-api` container via the docker-compose file.

## ℹ️ Context for reviewers
This PR adds a health check, via the `/v1/status` endpoint, in order to assert app containers, being automatically deployed as part of a PR, run healthy. The main purpose of this check is to cause a failure sooner in the deploy process should the container in question be deemed unhealthy.

## 👀 Demo
Running:
```
make release-test
```

Will poll check the health of the `mass-pfml-api` container, after starting a dev instance of it, and exits with a `1` if it's unresponsive/unhealthy beyond the configured number of retries. This is run on every PR branch.