CP-1104: fix fieldHasValue check (#1204)
🎫 Ticket
https://lwd.atlassian.net/browse/CP-1104

🛠 Changes
Added a "number" type check to fieldHasValue.

ℹ️ Context for reviewers
The current check for completed Steps in Step.js uses fieldHasValue to check for a value on the Claim. fieldHasValue uses an "isEmpty" check, but when the value is of a "number" type, "isEmpty" results in true. This fix updates the fieldHasValue helper to return true when the value is a "number" type.