End-254: E2E test for GBM1: Mailed HCP, MA Resident  (#1268)
https://lwd.atlassian.net/browse/END-254

This PR adds a new scenario: GBM1. This scenario is when a claimant submits a claim through the portal and then instead of uploading their HCP to the portal, they mail it in physically.

The CSR rep will then get a task for evidence review. For the purposes of current functionality - we are adding the "mailed" HCP form manually, then adjudication approval once evidence has been reviewed.

* Adding Files/Updating types/.features

* updating actions and cypress

* lint fixes

* change fineos function names

* remove console.log

Co-authored-by: Joseph Garcia <kingmoc@Josephs-MacBook-Pro.local>