END-310: Financial eligibility test (#1323)
https://lwd.atlassian.net/browse/END-310

This PR creates a test for financial eligibility. It also creates a new folder for this and similar smaller tests that only test part of the scenarios overall lifecycle.

What was previously named the Integration folder is now called "tests" and includes the "common" folder, as well as separate "e2e" and "integration" folders. Running "cypress open" in "tests" opens a Cypress window that shows all tests in subfolders.

* re-organize cypress file structure

* log to cypress instead

* linting

* fin elig test in progress

* cypress functions

* update path

* api updates

* change fineos path

* Remove claim approval from scenarios

* fix assertOnClaimPage function

* fix type in step name

* update JSON