END-317: E2E test for submitting part one of a claim (#1342)
This test asserts that you can commence intake on a claim in FINEOS, once it has been "part one" submitted in the portal. "Part one" means that everything is submitted except for evidence documents.

Currently, it seems that if you search for a random case number that does not correspond to a started application, FINEOS shows an error. However, if you search for a case number of a claim that has had Part One submitted, you are taken to a page that allows you to "commence intake." This test checks that this functionality is working as expected.