END-328: Fix HAP2 test (#1343)
Previously, this test covered the case where an OOS ID was expected, but not uploaded. However, a change in the portal meant that ID upload was required to submit an application fully. Accordingly, the portal flow for this test was changed to add document upload. However, the document being upload was the same "MA ID" placeholder document that was being uploaded for HAP1. This could potentially be confusing. Also, the test was failing because the old HAP2 test checked to confirm that no document was uploaded. The case where an invalid ID was submitted is now covered by UNH3. This test now approves the claim, which means that for the FINEOS side, it is identical to HAP1. It's not clear what additional business value this test provides, and we could consider removing it. The only difference between OOS and MA ID upload seems to be that a task should be created for manual ID verification for the former, but these e2e tests deliberately ignore task creation.