Remove NLB stickiness workaround (#1359)
Previously there was an "Error: Network Load Balancers do not support Stickiness" message if a stickiness setting was not provided. To work around that, we added a stickiness block that was disabled.

With v3.10.0 of terraform-aws this is no longer needed, and there is now also an error that "lb_cookie" is not a valid option for this load balancer.

Since we don't need stickiness anyways, remove it.