CP-1026: [Part 2] Add utilities to average hours across 7 days (#1384)
## 🎫 Ticket

https://lwd.atlassian.net/browse/cp-1026

## 🛠 Changes

- Adds `minutesWorkedPerWeek` param to `WorkPattern#addWeek` which will split the hours across a work week
- Adds `WorkPattern#minutesWorkedEachWeek` to recalculate average minutes after they are split across 7 days
- Adds utility to convert minutes to hours and minutes for API and User interface.

## ℹ️ Context for reviewers

There was a recent decision to collect hours worked per week and average across 7 days for both variable and rotating schedules. For rotating schedules, we will be collecting average hours _for each week that rotates_. We do not now nor do i think we should store the hours for each week in the API. Instead, this PR generates a work week from hours and adds a method to recalculate the hours from the week. Intended use is something like:

```js

<input 
    onChange={(hours) => {
      const updatedWorkPattern = WorkPattern.removeWeek(workPattern, 1);
      updateFields({ workPattern: WorkPattern.addWeek(updatedWorkPattern, hours * 60) });
    }
    value={convertMinutesToHours(workPattern.minutesWorkedEachWeek[0]).hours}
/>
```

🧐 
@lorenyu