API: Generate typecheck-able pydantic models for FINEOS Customer API (#1482)
con* annotations do not play nice with mypy[1], so use the
`--field-constraints` flag to the code generator to move the constraints
to the `Field` declaration instead of the type annotation, which
supports the same validation functionality as the con* annotation.

In addition, the transformation of the FINEOS money types to Decimal
types need to additionally drop the `maxLength` and `minLength`
constraints since they do not apply to Decimal types. So the
preprocessing of the Customer API specification needs to be slightly more
sophisticated than single line search-replace.

These changes enable us to stop ignoring the Customer API types in the
pydantic mypy checking. Which will enable more improvements to come.

This includes the manual edits to the models done in a recent commit[2].

Recreated `customer_api.py` with `make clean && make models`, then
added back in the manual edits.

[1] https://github.com/samuelcolvin/pydantic/issues/156
[2] a146c68ada267c3d1dab001c2d6c01d6288e5d4b