END: Cleanup unused typescript types (#1528)
Implements /claims endpoint which is a GET that takes a valid user ID and returns their first claim from the database.

This PR removes a slew of Typescript types that are no longer used. This should help prevent us from tripping over them or accidentally using them in the future. It also clears up an issue where the document tests were failing due to a console.warn, and fixes a flaky simulation generation test due to date comparisons.

* END: Cleanup unused typescript types

* Fix console.warn in document test, fix flaky simulation start date test