END-392: Improve password generation to always meet requirements (#1549)
https://lwd.atlassian.net/browse/END-392

This is an adjustment to the Cypress password generation to ensure that passwords always meet requirements. Previously, it was possible that passwords wouldn't have an uppercase or lowercase letter in them. Also, some of the currency symbols don't appear to be valid symbols, so we'll just switch to simpler ones.

* Demo: Unit tests and adhoc testing via swagger UI.