API-577: Lookup employer ID from Paid Leave DB for eligibility file creation (#1558)
Replaced the FINEOS API call to lookup the FINEOS employer id with use of the now stored database value employer.fineos_employer_id. The latter is stored in the Paid Leave DB by the process that loads and updates employers into FINEOS. If the value is not found in the DB we log an info statement and continue, similar to what we did when we did not find the employer in FINEOS.

Modified the eligibility feed test cases accordingly. Also had to modify the EmployerFactory artifact to set FINEOS customer number and FINEOS employer id for tests to work.