API-752: Create RDS subnet groups for new DBPrivate subnets (#1603)
https://lwd.atlassian.net/browse/API-752

Creates new subnet groups that point to the DBPrivate subnets for each VPC instead of the app private subnets.

ℹ️ Context for reviewers

We're moving our databases to new subnets (in the same VPCs). They should live in the DB-specific subnets instead of the app subnets.

In order to do that, we have to create new subnet groups that point to the DB subnets. In the interim, both subnet groups have to co-exist.