API-470: Create Service Agreement Using Service Agreement Endpoint (#1587)
Created function to create an employer, and if successful create a service agreement. Wrapped the method around some testing code to run locally. Creating this PR as draft to discuss whether creating employer/SA is the way we want to go or if we want separate jobs to create employers and then rescan to create service agreements. I think the former is the way to go

Tested with a DOR import and generated 200 employers and service agreements in DT2. FEINs range from 100000001 to 100000100.