CP-1222: Allow reduced schedule leave period hours/minutes fields to be empty (#1654)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-1222

## 🛠 Changes

- Update `InputHours` to allow an empty `value` prop
- Add `fallbackValue` argument to `getFunctionalInputProps` so you can override the default fallback value, which is an empty string. This was necessary since `InputHours` only expects a number when a `value` is set.
- Revert defaulting of `ReducedScheduleLeavePeriod` minute fields to `0`, changing back to default to `null`.

## ℹ️ Context for reviewers

Some discussion related to this [happened here](https://github.com/EOLWD/pfml/pull/1622#discussion_r516864219). The original outcome was that we would punt this to a [tech debt ticket](https://lwd.atlassian.net/browse/CP-1336), which is still relevant for the work pattern work. 

I went ahead and addressed the reduced leave schedule portion of this though, since I realized that these default values were getting overrwritten by the API response, which defaults these fields to `null`. If we wanted to continue to default to `0` on the Portal side, I would need to add logic to the `handleSave` callback on the `reduced-leave-schedule` page to set the fields to `0` if their values were `null`, which felt not ideal if the longer term plan was to change these defaults to `null`.