END: Add get-token command to retrieve a bearer token for given credentials (#1657)
Implements /claims endpoint which is a GET that takes a valid user ID and returns their first claim from the database.

Adds a CLI command to the E2E codebase to fetch a valid Bearer token given the current credentials (taken from environment). This is useful when doing Postman testing, which requires a bearer token. Also sets a limit the number of business simulation submission failures that can be received before aborting (this was necessary for Pilot 4 biz sim - just capturing it here).

* END: Add get-token command to retrieve a bearer token for given credentials

* Fix CS