API-789: Remove filter patterns for lambda subscriptions (#1709)
https://lwd.atlassian.net/browse/API-789

This removes the filter patterns for existing lambda log filters so that all logs are sent to New Relic.

Why: EOLWD is shifting away from Splunk and using New Relic Logs as the aggregation platform. Ada and I have confirmed with Shiva that they want us to be sending logs over to New Relic.