END-410: Use separate employees for each environment (#1740)
https://lwd.atlassian.net/browse/END-410

In END-410, we're going to be running Cypress against arbitrary environments. Unfortunately, we have different employees loaded into each environment, which makes our previous system of using the employees.json file untenable. This change adds JSON files generated during biz sim that can be used as a user pool. Cypress tests will then select a random employee from the pool that matches the given criteria. As a bonus, this should put the nail in the coffin of any lingering "too many claims for one employee" issues that we've been having.

This PR also updates the user account we use for Cypress testing in Github actions. Until we get the E2E_PORTAL_PASSWORD secret updated in Github actions, this will cause all Cypress tests to fail in Github actions. It's necessary that we do this though, because we want to use the same password in all environments.

Steps to test:

Run any Cypress test while pointed at test (E2E_ENVIRONMENT=test).
Run any Cypress test while pointed at stage (E2E_ENVIRONMENT=stage).
Verify that the employee was found in both environments (hopefully the claim should submit successfully, but we're not sure if that'll work or not).

* END-410: Use separate employees for each environment

* Fix CS

* Use fs.readFile instead of import()

* Use fs.readFile instead of import()