CP-1027: Use notification case id to fetch occupation, always send Fixed work pattern (#1752)
## 🎫 Ticket

Fixes for issues found during UAT for
https://lwd.atlassian.net/browse/cp-1027

## 🛠 Changes

- Uses notification id for fetching occupation record
- Sends all work patterns as `Fixed`

## ℹ️ Context for reviewers

#### Issue 1

Adding and updating work patterns was failing when a user had multiple claims. This was due to a misunderstanding of how FINEOS associates an employee's active Occupation with an Absence case.

Absence cases are child cases to Notification cases. While all cases have an Occupation record, only the Notification case is always associated with the active Occupation for the employee. The _first_ Absence case is associated with the active Occupation, but subsequent Absence cases get a new Occupation record that apparently cannot be associated with a WorkPattern.

#### Issue 2

FINEOS' API does not support `Variable` as a value for `workPatternType`. We will need to coordinate with CPS to determine if there is another way we need to record that the hours per day are averaged. 

## 🧐 

@lorenyu @jamesbursa